# docs/swagger.yaml
definitions:
...
  controller.CategoryDto:
    properties:
      name:
        maxLength: 64
        minLength: 2
        type: string
    required:
    - name
    type: object
...
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
      success:
        type: boolean
      timestamp:
        type: integer
    type: object
info:
  contact: {}
  description: 博客主要包含用户、文章和分类
  license:
    name: Apache 2.0
  title: Gin博客项目接口文档
  version: "1.0"
paths:
...
  /api/category/create:
    post:
      consumes:
      - application/json
      description: 添加分类，参数为分类name
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controller.CategoryDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 添加分类
      tags:
      - 分类接口
  /api/category/delete/{id}:
    delete:
      consumes:
      - application/json
      description: 通过分类id删除分类
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 删除分类
      tags:
      - 分类接口
  /api/category/detail/{id}:
    get:
      consumes:
      - application/json
      description: 通过分类id获取分类详情信息
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取分类详情
      tags:
      - 分类接口
  /api/category/disable/{id}:
    get:
      consumes:
      - application/json
      description: 通过分类id禁用分类
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 禁用分类
      tags:
      - 分类接口
  /api/category/enable/{id}:
    get:
      consumes:
      - application/json
      description: 通过分类id激活分类
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类id
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 激活分类
      tags:
      - 分类接口
  /api/category/list:
    get:
      consumes:
      - application/json
      description: 默认获取全部分类列表，分页参数为page_num和page_size
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 当前第X页
        in: query
        name: page_num
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 获取分类列表
      tags:
      - 分类接口
  /api/category/search:
    get:
      consumes:
      - application/json
      description: 通过分类name搜索分类列表
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类name
        in: query
        name: title
        type: string
      - description: 当前第X页
        in: query
        name: page_num
        type: integer
      - description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 搜索分类
      tags:
      - 分类接口
  /api/category/update/{id}:
    put:
      consumes:
      - application/json
      description: 更新分类，通过分类id找到分类并更新
      parameters:
      - description: Bearer 用户令牌
        in: header
        name: Authorization
        required: true
        type: string
      - description: 分类id
        in: path
        name: id
        required: true
        type: integer
      - description: 分类参数
        in: body
        name: object
        required: true
        schema:
          $ref: '#/definitions/controller.CategoryDto'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: 更新分类
      tags:
      - 分类接口
...
swagger: "2.0"